+++ packages/kotlin-markdown-core/src/jniMain/kotlin/com/nouprax/markdown/core/wire/JniMarkupDecoder.kt
@@ -143,6 +143,10 @@ private class JniTreeDecoder(
                 readChildren { consume(Mark(it, scope, anchor, attributes)) }
             }
 
+            JniNodeKind.INSERTION -> {
+                readChildren { consume(Insertion(it, scope, anchor, attributes)) }
+            }
+
             JniNodeKind.LINK -> {
                 val resource = resource()
                 readChildren { consume(Link(resource.first, resource.second, it, scope, anchor, attributes)) }