+++ packages/markdown-core/core/node.c
@@ -39,6 +39,7 @@ bool markdown_core_node_can_contain_type(markdown_core_node *node, markdown_core
     case MARKDOWN_CORE_NODE_EMPHASIS:
     case MARKDOWN_CORE_NODE_STRONG:
     case MARKDOWN_CORE_NODE_MARK:
+    case MARKDOWN_CORE_NODE_INSERTION:
     case MARKDOWN_CORE_NODE_LINK:
     case MARKDOWN_CORE_NODE_MEDIA:
         return MARKDOWN_CORE_NODE_TYPE_INLINE_P(child_type);
@@ -436,6 +437,8 @@ const char *markdown_core_node_get_type_string(markdown_core_node *node) {
         return "strong";
     case MARKDOWN_CORE_NODE_MARK:
         return "mark";
+    case MARKDOWN_CORE_NODE_INSERTION:
+        return "insertion";
     case MARKDOWN_CORE_NODE_LINK:
         return "link";
     case MARKDOWN_CORE_NODE_MEDIA: